minor vampire buff (youngling VL's rejoice) (#3139)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

-So after infinite stamina was killed vampires were left with all their
weaknesses (they instantly crumble to silver and miracles)
This PR gives them master athletics and expert unarmed with a stat boost
of 1 or 2 STR depending on your luck and +1 SPD (tell me in disc or in
comments if you want me to tweak the numbers)

-Age stats don't count for vampires and maniacs anymore. (you get the
same stats as an adult would)


## Why It's Good For The Game
-You were pretty fucked if you rolled youngling VL or maniac. (you
shouldn't)

-Unarmed skill for non lethal takedowns.

-Makes sense for vampires to be slightly stronger and faster than their
mortal counterparts.

-Why master athletics? Check video below. (Vampires should have good
stamina and frontflips are peak)


https://github.com/user-attachments/assets/c0c7dd03-66e7-4167-9ae7-7d76cfc07031


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
balance: Minor vampire buff 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->