Vampire Misc Changes (#3132)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Nosferatu Rat Transformation is a small rat instead of the big rat
with the reduced cooldown from 50 to 30.

- Nosferatu now has the Keen Ears trait.

- The pre-rework Vampire Telepathy is back with a few tweaks, now it
uses 25 Vitae and has a 30 seconds cooldown.

- Vampires assigned to positions that can assign other vampires can now
contribute to Vampire Lord Projects, they can finish projects such as
plate and amulet but they can't finish the Vampire Lord rites since a VL
is needed there to gain the reward.

- Vampire Lord now has the trait of foreigner even if they didn't had
them already.

- Vampire Lords does not get the Lady or Lord forced title anymore

- Fix Stealing the Sun and Making Death Knights not using vitae

- Fix Stealing the Sun giving free vitae to the VL

Closes: #3180 

## Why It's Good For The Game

- Big Rat was useless, even tho it assigns "15 Speed", mob speed is very
slow for some reason, a big target and everyone goes for it, a small rat
fits them and it is way easier to conceal yourself as it.

- For a Clan that is supposed to collect information and hide from
people the keen ears fits them well so they can act more covert than
ever.

- Telepathy was needed back to the Vampires, Zizo Cultists & Werewolves
already got a similar mechanic and to Vampires must act together for
greater results, while the Subordinates Messages worked it was only a
one way conversation, now they can't just spam telepath with the
cooldown and it has a small price to do it.

- These same Vampires already could sire new vampires, i don't see why
not make so they can contribute to the VL Projects, this will hopefully
increase the likelyhood of VL siring new people instead of going into a
one man army type of gameplay.

- This is to prevent people metagaming since you could exam someone with
the Lord Prefix and since they are not foreigner it could just appear :
"This is Lord John.", without any title or being foreigner so they were
not a Noble because there was no title nor a pilgrim Noble since they
are not a foreigner, even with the change below, this would prevent VLs
being not foreigners and having no titles. (You could argue since it
took so long for them to wake up people don't know the VL and consider
them a foreigner)

- I am still not sure about the removal of the Lord/Lady forced name,
but i think that this will give more liberty to the VL players they
could still be a Lord if they wish, or a Sir/Dame or anything at all,
tho if people dislike this change i am up to removing it.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
bal: Nosferatu transformation now turns them into a small rat instead of
the big rat.
bal: Nosferatu Clan gains the Keen Ears trait
bal: Vampires have their telepathy back, slightly modified.
add: Vampires with a position that enables them to assign other vampires
can now contribute to the VL Projects.
add: The Vampire Lord is a foreigner
del: The VL Lord/Lady auto names title is removed.
fix: Death Knight and Steal the Sun now uses the VL vitae
fix: Steal stun giving free vitae to the VL 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>