Assassin Fixes and Tweak (#3169)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Assassin is now average in medicine.

- Assassin can't steal someone identity if the body does not have a
head.

- Fixes Assassins not having the title of jobs like the inquisitor

- Fixes Assassins losing their limb if their copied identity didn't had
a limb

- Changed from TRAIT_PROFANE to TRAIT_GENERIC, didn't found anything
about what profane was supposed to do and it stopped working for me
after some testings the target would not be disfigured anymore.

Closes: #3162 

## Why It's Good For The Game

- Without average medicine they can't attach a head to a body.

- Even tho it is magical, how can you steal someone identity if you
don't even know whose body is it in the first place.

- Fix!

- This don't fixes assassins being able to run without a leg after
copying someone body and have their log chopped off, i don't know why it
happens and i could not figure it out, it was like the leg was not
attached to the mob even tho it was, so even if you choped it they can
run like everything is okay (They still get hit with the blood loss like
usual)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
bal: Assassins got average medicine skill
bal: Assassins can't steal the identity of a headless body
fix: Assassins don't lose their limbs if their copied identity body
didn't had them
fix: Assassins not getting the title of some jobs that they stole the
identity of.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>