Steam knight armour slop (#3116)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Sort of a follow up PR to #3115 , given how that PR gives steam knights
the engineering skill.
New artificer crafting category, armor.
You can craft steamknight armour, requires master in engineering. Is, I
hope, expensive enough.
Steamknight armour now requires skilled in engineering to be powered on.
Steamknight boiler now requires a backpack to craft.

Also a minor tweak to the cooling backpack recipe name, so its inline
with the other recipe names.

## Why It's Good For The Game

Most, if not almost all, things are craftable in the game. I vaguely
recall steamknight armour was meant to be craftable (and was for a short
period) but was removed due to balance reasons (I think? Not entirely
sure)
Anyways, the armour is now craftable. Expensive enough to that getting
it will be an expensive feat.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Steamknight armour is now craftable
balance: Steamknight armour now requires skilled in engineering to be
properly utilized
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->