new cloak + mask resprite (#3124)
## About The Pull Request

a player set up a bounty to redrip the adepts in masks and a special
cloak for the reformed thief, here are the sprites and code

adds the code for the new cloak + the resprited silver mask, the cloak
has a movement for onmobs, sadly doesn't work for some reason



https://github.com/user-attachments/assets/f2b82531-395d-42a4-9c62-f9bc906df288



## Why It's Good For The Game

is a different way to see the adepts, their new masks make them more
epic to see

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->


:cl:
add: Added more things
del: Removed old things
fix: fixed a few things
image: added/modified/removed some icons or images
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.