Wicker handbaskets (#3195)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds a new, craftable storage item called a wicker handbasket. It costs
3 fiber to slapcraft and has a capacity of 9. It is essentially a worse
sack but doesn't require any sewing to craft. Can be worn on the hip.
Has 5 unique overlay sprites dependent on items stored within. The
handbasket must be holding at least 5 of the item (a majority of the
storage) to show the overlay. Bread, herbs, fruit, and eggs are
currently represented as well as a generic overlay for anything else.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More unique storage options for flavor and convenience.

![ss+(2025-09-11+at+03 43
54)](https://github.com/user-attachments/assets/fc9b16fe-e903-439d-bda0-c8c462763e6e)
Eggs, fruit, herbs, bread, and generic sprites. Empty at the top.


https://github.com/user-attachments/assets/f33452eb-64be-40ec-9301-67364775401e


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:Bloo4420
add: Craftable wicker handbaskets with unique content sprites
code: changed 3 instances of (things.len) in bags.dm to instead be
(length(things)) to be in line with other code.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>