Map changes (#3340)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Title
<img width="590" height="592" alt="image"
src="https://github.com/user-attachments/assets/1101de42-81eb-4dcc-9de6-382855313eb5"
/>
Fixes this... thing
and rotates the bench near the fountain in the city
<img width="816" height="650" alt="image"
src="https://github.com/user-attachments/assets/43763d68-72f8-4127-bb07-209098b46d93"
/>
Because it was facing south
and I didn't like that, because you had to walk through the bushes to be
able to sit on it

## Why It's Good For The Game

Basically what I said in the "about the pull request", so you can use
the bench

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
map: modified the vanderlin map a bit, honestly you won't notice
anything
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->