Language and Lore Updates (#2637)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Did you know there's still a ton of stuff that's either missing
descriptions, using old lore, *lore breaking* lore, or just outdated
language?

Or that the Stewards outfit was victorian (Who the fuck is that?!!! She
sounds like a nerd)

Or that Zybantu was still referenced.

Or that plums, lemons, limes, tangerines, and pears all had the same
description?

I know I did!

This PR changes a bunch of that.  

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Corrects incosistencies, fixes old shit, updates language, providing for
a more coherent idea of both the lore and themes.

Aberra wants this.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>