Adds Lakkari's lore to the Lore Primer (#3377)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR is meant to add the Queendom of Lakkari (alongside its cities)
to the server's lore primer.

## Why It's Good For The Game

I believe Africa is often misrepresented or underrepresented in fantasy
settings. I wished to remedy that with Lakkari - a nation inspired by
Egypt and Sudan. I have spent months developing Lakkari with a private
lore doc with the feedback of staff members such as Aberra (shoutout to
her btw), which I used to develop the lore contained within this pr.

## Changelog

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Lakkari and Lakkarian cities lore to the lore primer
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->