Steamknight Armor Toggle Refactor (#3253)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Scrap the old system but of course keeping the signals, now instead of
auto activating it, you need to right click the boiler while you have
the full set of steamknight armor to turn it on, pass the enginner skill
check, then it has a delay based of the player enginner skill.

If you remove the boiler or part of the armor without turning off you
will lose a rng amount of steam between 100-200 so turn off your armor.

You can't charge steam items that are broken, at the moment only the
boiler can break if the user keeps hitting it with their sword. (Drill
auto destroy itself and the harpoon can't be hit)

## Why It's Good For The Game

Even tho it lacks sprite, i believe this is way more simple both in
terms of code and gameplay, right click the boiler to both turn it on
and off, it also makes the code way more cleaner and easier to
understand, now it is prob possible to put the unpowered and powered FOV
but i will leave that to another PR.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
refactor: How the Steamknight Armor turns off and on
qol: To turn the steamknight armor on or off you can right click the
boiler
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>