Changes to how ingot moulds work (#3133)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

When ingot molds finish hardening the metal into an ingot, they don't
need to cool the metal down again.
Naturally, this makes no sense, and turned out it was a bug.
This PR fixes this bug, but also quickens the cooling process.
Also adds descriptions to the crucible and moulds.
Also, made a new contraption that cools down metals into ingots even
faster. Doesn't need to be an artificier contraption, more so a "proof
of concept".

## Why It's Good For The Game

I... I don't know
I mean all this does in practice is make it so smiths cant plop out
ingots like mad
It's why I made the cooling process faster, so people can still make
ingots without having to wait years
I guess I found it really jarring how you have to wait for the ingot
mold to harden the metal, but then don't have to wait after it hardens
the first time

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Added a new artificer contraption, basically an ingot mold that
cools down metal twice as fast as a normal one (normal ingot moulds cool
down THRICE as fast as before now)
fix: fixed ingot molds not needing to reharden metal after the first
harden-ing
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->