Woodcutter's Hut - Minor Forest Addition (#3402)
## About The Pull Request

Adds a Woodcutter's Hut to the Forest, with rats inhabiting inside. Some
basic loot is scattered here or there, should be enough for the level of
danger that lingers there. Found deeper in the forest to the South West
side. Do let me know if there are any changes people want in-place, this
is my first PR for Vanderlin.

<img width="672" height="1216" alt="image"
src="https://github.com/user-attachments/assets/5123e8cf-cfff-4482-b8dc-a8f21b82b3fe"
/>

## Why It's Good For The Game

More unique areas for players to explore in the Forest, this one is a
bit more hidden compared to some others. Adds variety to the area.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
map: Adds new Woodcutter's Hut map feature to the Forest
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->