Poison Resilience and minor changes (#3190)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds poison resilience trait to dwarves, which is basically what they
already had, a resilience to certain poisons
they are also now resilient to (strong) stamina poison (but not killers
ice!)
also did some tweaks to the way graggar boon interacts with eating
organs, tl;dr instead of the reagent giving the foodbuff the, foodbuff
instead is changed if the player has the boon (tested and it works)

Also buffed eating livers because, well they are eaten irl same as
hearts and lungs. iirc other organs aren't.

no changelog because there's barely anthing noteworthy

## Why It's Good For The Game

Makes the code cleaner I believe, and poisons dont check for a specific
race but a trait instead

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>