Changes the combat mode music for Servant roles (#3391)
## About The Pull Request

This PR changes the Servant role and it's adjacent classes (Tapster,
Ring servant)'s combat modes to the following tracks :

- The Tapster uses Hellion, which is the track used for all of the inn
(and the butler for some reason).
https://www.youtube.com/watch?v=7z31EC8umrQ
- The Servant uses Amontillado, shared with the prisoner class. This is
because a) The track sees very little use (prisoners are rarely played)
and b) the servants are indentured to hell and back and figuratively
prisoners of the keep. https://www.youtube.com/watch?v=R30aVqcrEUU
- The Ring Servants gets "The Reckoning", this is normally a track used
by the Aasimar mercenary subclass, but I feel like it fits the ring
servant, as they're somewhat implied to outlast their assigned gaffer,
somewhat like how aasimar outlast others.
https://www.youtube.com/watch?v=xrXT1x8N4PU

## Why It's Good For The Game

These jobs used the old, very underwhelming combat mode music, this
changes that to the usual bops by T87-Sulfurhead.

## Changelog

:cl:
sound: Changed the combat mode music for the Servant, Tapster and Ring
Servant roles.

/:cl:

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

I have a video file showcasing the 3 cmodes on a test server, ask me for
it on discord if you want.

Co-authored-by: Somniworld <pablo.capel4@orange.fr>