Forest Chapel - Minor Forest Addition (#3408)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds an old abandoned chapel to the Forest, which has fallen to heavy
decay and ruin. Plants grow all over the area, walls crumble and fences
fall. However, a pair of goblins have set up in the ruins for safety,
which players may encounter.

<img width="800" height="1248" alt="image"
src="https://github.com/user-attachments/assets/cbe0d5dd-4560-45f6-a8b4-06b4d9c8e97f"
/>

## Why It's Good For The Game

The older cemetery to the NE was a bit lackluster, so I updated it, gave
it a whole chapel, some enemies sprinkled in, etc. Just adds some
variation so that players can get a better experience

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
map: Added the Forest Chapel to the North East of the map.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->